MRD-2609 Support 'Sentenced Under' field setting (#259)
Support for setting a Sentence's 'Sentenced Under' field is added here. The new
field is mandatory, in line with how the 'Released Under' field for Releases is
mandatory, as both these fields should refer to the same legislation when the
Release falls under the Sentence.

The new sentencedUnder field is included here in the SentenceComparator's
logic. Although the reason behind the exclusion of some of the Sentence's
fields from the comparison is unclear, the new field was included, as:
 - Most of the existing fields have been included so far.
 - We see no reason why two sentences with different sentencedUnder values
    should be considered to be the same.