Remove explicit values from ReferenceTest (#272)
The tests for the reference list endpoints are checking whether the returned
lists have specific values, which makes them brittle against changes to the
values lists on the PPUD side. Additionally, as has recently happened, the PPUD
team first adds a new value to a list in the Dev environment for testing
before they decide to incorporate it into the Prod environment. This means Dev
will be out of sync with Pre-Prod (and Prod), meaning we could be stuck without
being able to get changes to our PPUD Automation code through both
environments (e.g. if the last value of a list is checked in a test, but a new
value is added going after that; if we update the test before the change in
PPUD goes into Prod, the test would pass in Dev but not in Pre-Prod).

To avoid this, and since we aren't really concerned with what specific values
each list returns (with the exception that we check the special 'Not Specified'
value isn't returned, and that for custody-type we only return the 'Determined'
value), the tests are here updated to simply check that the affected endpoints
return at least one value and don't include the 'Not Specified' value.