MRD-2712 Add OpenApiDocsTest
Following on a conversation regarding an upgrade after which the Swagger UI
page stopped working in another project, it was highlighted that an automated
test can and should be set up to ensure this doesn't happen when changes are
made. The test added here is based on the one in
hmpps-prisoner-from-nomis-migration.

It is unclear at the moment whether the bearer token part is needed, so the
test has been commented out and will be either uncommented or removed in a
follow-up.