MRD-2684 Update releasedUnder on indeterminates
The releasedUnder field of a Release for an Indeterminate sentence should only
be set to one of two values (IPP Licence or Life Licence, depending on the
specific custody type). The Release update code is changed here accordingly.

Note that the changes don't affect the only determinate case currently
supported, as releasedUnder continues to be user-set in such cases.