MRD-2768 Upgrade dependencies (#291)
Non-breaking dependency upgrades have been applied here.

The bouncycastle dependency was pinned a while back to address a CVE. The
dependency originally came from webdrivermanager, which has since removed
bouncycastle as one of its dependencies, so we remove ours too.

The io-commons dependency no longer needs to be pinned, as the webdrivermanager
dependency we transitively get it from has stopped pulling the version with a
CVE.