MRD-2909 Control police-forces retrieval with feature flag
The retrieval of the police-force list of values from PPUD was updated recently
to factor in changes made to the relevant page in PPUD. Although this fixed our
code in dev, it broke in preprod and prod, as PPUD stage their releases, with
each release including a group of changes. To handle this, we are adding here a
new feature flag to determine how to retrieve said list. Technically, this is
no longer needed, as the changes in PPUD have been rolled out to preprod and
prod and we are once again aligned on all three environments, but we are still
adding the flag to test we are correctly setting up the framework in all three
environments (PPDU Automation isn't yet live, so we can afford to do this).