MRD-2778 Fix indet. sentence rel. date retrieval (#297)
The release date for indeterminate sentences was being retrieved from the wrong
place: the D.O.R. field contains the Date of Remand, not the Date of Release.
The correct way of retrieving it is to take the latest date from the Release
records linked to the relevant sentence, as corrected with these changes.