MRD-2684 Update indeterminate sentence releases
The code is here updated to support updating the release of an indeterminate
sentence. As opposed to the determinate case, we should never be adding a new
release to the sentence, but updating the latest one. The set of fields that
need to be updated is also slightly different compared to the determinate case:
we are not trying to match on existing releases, so the key fields can be
updated and should not be ignored.

We were also provided with busines logic to determine what the 'released under'
field should be set to based on the custody type in indeterminate cases,
implemented here through the calculateReleasedUnder method, with its result, if
not null, being used instead of the releasedUnder value provided with the
update request (as things stand, we don't expect to receive a releasedUnder
value in a request to update an indeterminate sentence, but the code does allow
it and would also cover future indeterminate cases where the request value
should be used if the custody type is configured so).