MRD-2684 Support updating indeterminate sentences (#285)
Support for booking recalls onto PPUD for existing indeterminate sentences is
added here. This includes fixing the value specific fields (recall type, licence
type...) are set to based on the relevant sentence's custody type.

A refactor of some of the internal workings of the code has also been carried
out as part of this development, establishing a pattern for splitting out code
from the OperationalPpudClient class, which had become a god class.

Updated to also compare indeterminate sentences:
I realised too late that this wasn't needed anymore, since we only need to
support updates of indeterminate sentences, where the comparison isn't done.
Since I had already written the code and we will likely eventually support
creating indeterminate sentences (where the comparison code will be relevant to
prevent duplicates), I thought it better to keep these changes in.

The code is here updated to support updating the release of an indeterminate
sentence. As opposed to the determinate case, we should never be adding a new
release to the sentence, but updating the latest one. The set of fields that
need to be updated is also slightly different compared to the determinate case:
we are not trying to match on existing releases, so the key fields can be
updated and should not be ignored.