Correct integration test recall dates (#311)
PPUD was recently updated to check whether the following dates set in the
recall screen are set in strict ascending order (all refer to the Actual
column on the screen, where the user can set the dates):
 - Decision Following Breach
 - Received by Section On
 - Recommended to/on
 - Issue revocation
 - Return To Custody

We set the first three, but in some of our tests we are setting them to
random values. This is adjusted to ensure the ordering is satisfied.

Note that PPUD does allow them not to be in strictly ascending order, but a
modal window is opened each time this is done to alert the user and ask to them
to confirm they want to do so. A separate ticket may be opened to look into
whether we need to support unordered dates and, if so, to adjust the code
accordingly.