MRD-2780 Migrate pipeline to GHA
Our pipeline is here migrated from CircleCI to GitHub Actions.

Three other sets of changes are happening here due to circumstances linked to
the migration:
 - Redis is turned off for integration tests.
 - We unpin the Firefox version for tests
 - SonarQube and jacoco are removed.

Redis removal for integration tests
Mocking away the browser showing the PPUD pages is unfeasible, so our
integration tests need to run against a real instance of PPUD (their test one).
The PPUD team have had issues with unknown entities probing their security, so
they have set up an IP address allow-list. Our CircleCI set-up was running
from IP addresses known to them, but GitHub has a much wider and dynamic range
of IPs they can't support. Instead, we use HMPPS self-hosted runners. However,
these runners have restrictions which prevent us from easily (if at all)
running Redis, so we turn it off just for the integration tests.

Unpinning Firefox version for tests
The existing CircleCI code had pinned the Firefox version to install to what
the latest ESR version was at the time. This only affected the pipeline
integration testing, not the production set-up (where the latest ESR at
installation time is installed). This was originally going to be addressed in
MRD-2697, but I ran into issues installing the pinned version in the pipeline,
so I have remove the pinning altogether. This means that the latest ESR
version of Firefox will be installed in the pipeline, matching production.

SonarQube and jacoco removal
SonarQube is already due to lose support at some point, as was announced in
Slack a while back, so we've decided to use this migration to remove it
altogether. With this, we also remove jacoco, which was generating the data for
SonarQube. We haven't been doing separate checks on the code coverage data
jacoco produces, so there's no point in keeping it. If we do need it in the
future we can always re-add it.